Move assembly routines to own file
